Moved Firebase Service to a separate folder
